Refactor final

+++ EmqxLearning.Shared/Concurrency/Configurations/RateLimiterOptions.cs
@@ -3,5 +3,5 @@ namespace EmqxLearning.Shared.Concurrency.Configurations;
 public class RateLimiterOptions
 {
     public string Name { get; set; }
-    public long InitialLimit { get; set; }
+    public int InitialLimit { get; set; }
 }
